requirements.txt
chromadb>=0.3.21  # Replace with the required version
gradio
langchain_community
langchain_core
langchain_groq
langchain_chroma
pandas
PyPDF2
python-dotenv